commit to save progress before massive rehaul
