+++ C_code/dcel/dcel.h
@@ -1,7 +1,7 @@
 #ifndef h_dcel
 #define h_dcel
 /* TODO: create linked list for vertex to store list of incidentedge for faster queering --DONE--*/
-
+/* TODO: change the liked list structure for mergesort */
 
 /*          Structure           */
 
@@ -12,8 +12,8 @@ struct dcel_vertex {
   double y;
   struct node *head;/* head of the linked list */
   struct node *z; /* end of the linked list */
-  int top = 0;
-  struct dcel_edge *edge_array /* pointer to an array of edge */
+  int top; /* postion to put in the edge array */
+  struct dcel_edge **edge_array; /* pointer to an array of edge */
 };
 
 struct dcel_edge {
@@ -35,6 +35,7 @@ struct point { /* structure of the 2 points */
 
 struct node { /* structure of a linked list node */
   struct dcel_edge *half_edge;
+  double angle; /* temporary angle use for the sorting step */
   struct node *next;
 };
 
@@ -44,8 +45,7 @@ struct node { /* structure of a linked list node */
 
 struct dcel_vertex *new_vertex(double x, double y);
 struct dcel_edge *new_edge(struct dcel_vertex *org, struct dcel_vertex *des);
-struct dcel_edge *new_polygon_edge(struct dcel_vertex *org, struct dcel_vertex *des);
-struct dcel_face *new_triangle(struct dcel_vertex *p1, struct dcel_vertex *p2, struct dcel_vertex *p3);
+/*struct dcel_face *new_triangle(struct dcel_vertex *p1, struct dcel_vertex *p2, struct dcel_vertex *p3);*/
 /*struct dcel_vertex *next_vertex(struct dcel_vertex *vertex);, you should not find other vertex through a vertex, they only connected by edge*/
 void next_prev(struct dcel_edge *first, struct dcel_edge *next);
