+++ C_code/dcel/make_dcel.h
@@ -1,0 +1,13 @@
+#ifndef make_dcel_h
+#define make_dcel_h
+
+
+/*          Function prototype          */
+
+int read_polygon_file(char *filename);
+int read_edge_file(char *filename);
+void create_dcel(void);
+
+#endif /* ifndef make_dcel_h */
+
+
