+++ C_code/dcel/sample_edge_file
@@ -1,0 +1,2 @@
+(1, 5) (5, 2) (2, 4)
+
