+++ C_code/dcel/sample_polygon_file
@@ -1,1 +1,1 @@
-(0, 0) (5, 1) (6, 4) (5, 6) (2, 4)
+(0, 0) (5, 1) (6, 4) (7,6) (5, 6) (2, 4)
