--- C_code/dcel/sample_triangle_file
@@ -1,4 +1,0 @@
-(0, 0) (5, 1) (2, 4)
-(5, 1) (6, 4) (2, 4)
-(2, 4) (6, 4) (5, 6)
-
