+++ C_code/dcel/scan.h
@@ -1,0 +1,8 @@
+#ifndef scan_h
+#define scan_h
+
+/*         functions prototype         */
+struct dcel_vertex *scan_point(FILE *pos);
+struct dcel_edge *scan_edge(FILE *pos);
+
+#endif /* ifndef scan_h */
